Add new GenShell to UI project
