+++ code/src/UI/Services/ProjectConfigInfoService.cs
@@ -6,7 +6,7 @@ using System;
 using System.IO;
 using System.Linq;
 using Microsoft.Templates.Core;
-using Microsoft.Templates.Core.Gen;
+using Microsoft.Templates.Core.Gen.Shell;
 using Microsoft.Templates.Core.Helpers;
 using Microsoft.Templates.Core.Resources;
 using Microsoft.Templates.Core.Services;
@@ -31,16 +31,16 @@ namespace Microsoft.Templates.UI.Services
         private const string MenuBar = "MenuBar";
         private const string Ribbon = "Ribbon";
 
-        private readonly GenShell _shell;
+        private readonly IGenShell _shell;
 
-        public ProjectConfigInfoService(GenShell shell)
+        public ProjectConfigInfoService(IGenShell shell)
         {
             _shell = shell;
         }
 
         public ProjectMetadata ReadProjectConfiguration()
         {
-            var projectPath = _shell.GetActiveProjectPath();
+            var projectPath = _shell.Project.GetActiveProjectPath();
             var projectMetadata = ProjectMetadataService.GetProjectMetadata(projectPath);
 
             if (IsValid(projectMetadata))
@@ -120,7 +120,7 @@ namespace Microsoft.Templates.UI.Services
 
         private bool IsUwp()
         {
-            var projectTypeGuids = _shell.GetActiveProjectTypeGuids();
+            var projectTypeGuids = _shell.Project.GetActiveProjectTypeGuids();
 
             if (projectTypeGuids != null && projectTypeGuids.ToUpperInvariant().Split(';').Contains("{A5A43C5B-DE2A-4C0C-9213-0A381AF9435A}"))
             {
@@ -334,7 +334,7 @@ namespace Microsoft.Templates.UI.Services
                     {
                         if (ExistsFileInProjectPath("Services", "ActivationService.cs"))
                         {
-                            var codebehindFile = Directory.GetFiles(Path.Combine(_shell.GetActiveProjectPath(), "Views"), "*.xaml.cs", SearchOption.TopDirectoryOnly).FirstOrDefault();
+                            var codebehindFile = Directory.GetFiles(Path.Combine(_shell.Project.GetActiveProjectPath(), "Views"), "*.xaml.cs", SearchOption.TopDirectoryOnly).FirstOrDefault();
                             if (!string.IsNullOrEmpty(codebehindFile))
                             {
                                 var fileContent = File.ReadAllText(codebehindFile);
@@ -346,7 +346,7 @@ namespace Microsoft.Templates.UI.Services
 
                     if (ExistsFileInProjectPath("Services", "ActivationService.vb"))
                     {
-                        var codebehindFile = Directory.GetFiles(Path.Combine(_shell.GetActiveProjectPath(), "Views"), "*.xaml.vb", SearchOption.TopDirectoryOnly).FirstOrDefault();
+                        var codebehindFile = Directory.GetFiles(Path.Combine(_shell.Project.GetActiveProjectPath(), "Views"), "*.xaml.vb", SearchOption.TopDirectoryOnly).FirstOrDefault();
                         if (!string.IsNullOrEmpty(codebehindFile))
                         {
                             var fileContent = File.ReadAllText(codebehindFile);
@@ -359,7 +359,7 @@ namespace Microsoft.Templates.UI.Services
                 case Platforms.Wpf:
                     if (ExistsFileInProjectPath("Services", "ApplicationHostService.cs"))
                     {
-                        var codebehindFile = Directory.GetFiles(Path.Combine(_shell.GetActiveProjectPath(), "Views"), "*.xaml.cs", SearchOption.TopDirectoryOnly).FirstOrDefault();
+                        var codebehindFile = Directory.GetFiles(Path.Combine(_shell.Project.GetActiveProjectPath(), "Views"), "*.xaml.cs", SearchOption.TopDirectoryOnly).FirstOrDefault();
                         if (!string.IsNullOrEmpty(codebehindFile))
                         {
                             var fileContent = File.ReadAllText(codebehindFile);
@@ -380,7 +380,7 @@ namespace Microsoft.Templates.UI.Services
             {
                 if (IsCSharpProject())
                 {
-                    var codebehindFile = Directory.GetFiles(Path.Combine(_shell.GetActiveProjectPath()), "*.xaml.cs", SearchOption.AllDirectories).FirstOrDefault();
+                    var codebehindFile = Directory.GetFiles(Path.Combine(_shell.Project.GetActiveProjectPath()), "*.xaml.cs", SearchOption.AllDirectories).FirstOrDefault();
                     if (!string.IsNullOrEmpty(codebehindFile))
                     {
                         var fileContent = File.ReadAllText(codebehindFile);
@@ -432,17 +432,17 @@ namespace Microsoft.Templates.UI.Services
 
         private bool IsCSharpProject()
         {
-            return Directory.GetFiles(_shell.GetActiveProjectPath(), "*.csproj", SearchOption.TopDirectoryOnly).Any();
+            return Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.csproj", SearchOption.TopDirectoryOnly).Any();
         }
 
         private bool IsVisualBasicProject()
         {
-            return Directory.GetFiles(_shell.GetActiveProjectPath(), "*.vbproj", SearchOption.TopDirectoryOnly).Any();
+            return Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.vbproj", SearchOption.TopDirectoryOnly).Any();
         }
 
         private bool IsCppProject()
         {
-            return Directory.GetFiles(_shell.GetActiveProjectPath(), "*.vcxproj", SearchOption.TopDirectoryOnly).Any();
+            return Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.vcxproj", SearchOption.TopDirectoryOnly).Any();
         }
 
         public string GetProgrammingLanguage()
@@ -467,7 +467,7 @@ namespace Microsoft.Templates.UI.Services
         {
             try
             {
-                return Directory.GetFiles(Path.Combine(_shell.GetActiveProjectPath(), subPath), fileName, SearchOption.TopDirectoryOnly).Count() > 0;
+                return Directory.GetFiles(Path.Combine(_shell.Project.GetActiveProjectPath(), subPath), fileName, SearchOption.TopDirectoryOnly).Count() > 0;
             }
             catch (DirectoryNotFoundException)
             {
@@ -483,7 +483,7 @@ namespace Microsoft.Templates.UI.Services
         {
             try
             {
-                var filePath = Path.Combine(_shell.GetActiveProjectPath(), subPath, fileName);
+                var filePath = Path.Combine(_shell.Project.GetActiveProjectPath(), subPath, fileName);
                 var fileContent = FileHelper.GetFileContent(filePath);
                 return fileContent != null && fileContent.Contains(lineToFind);
             }
@@ -499,7 +499,7 @@ namespace Microsoft.Templates.UI.Services
 
         private bool ContainsNugetPackage(string packageId)
         {
-            var projfiles = Directory.GetFiles(_shell.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
+            var projfiles = Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
             foreach (string file in projfiles)
             {
                 if (File.ReadAllText(file).IndexOf($"<packagereference include=\"{packageId}", StringComparison.OrdinalIgnoreCase) != -1)
@@ -508,7 +508,7 @@ namespace Microsoft.Templates.UI.Services
                 }
             }
 
-            var configfiles = Directory.GetFiles(_shell.GetActiveProjectPath(), "packages.config", SearchOption.TopDirectoryOnly);
+            var configfiles = Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "packages.config", SearchOption.TopDirectoryOnly);
             foreach (string file in configfiles)
             {
                 if (File.ReadAllText(file).IndexOf($"<package id=\"{packageId}", StringComparison.OrdinalIgnoreCase) != -1)
@@ -522,7 +522,7 @@ namespace Microsoft.Templates.UI.Services
 
         private bool ContainsSDK(string sdkId)
         {
-            var files = Directory.GetFiles(_shell.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
+            var files = Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
             foreach (string file in files)
             {
                 if (File.ReadAllText(file).IndexOf($"sdk=\"{sdkId}\"", StringComparison.OrdinalIgnoreCase) != -1)
@@ -536,7 +536,7 @@ namespace Microsoft.Templates.UI.Services
 
         private bool AppContainerApplication(string property, string value)
         {
-            var files = Directory.GetFiles(_shell.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
+            var files = Directory.GetFiles(_shell.Project.GetActiveProjectPath(), "*.*proj", SearchOption.TopDirectoryOnly);
             foreach (string file in files)
             {
                 if (File.ReadAllText(file).Contains($"<{property}>{value}</{property}>"))
