+++ code/src/UI/Validators/HasNoCoreProjectValidator.cs
@@ -49,10 +49,10 @@ namespace Microsoft.Templates.UI.Validators
 
         private bool HasCoreProject()
         {
-            var extension = (GenContext.ToolBox.Shell.GetActiveProjectLanguage() == ProgrammingLanguages.CSharp) ? "csproj" : "vbproj";
+            var extension = (GenContext.ToolBox.Shell.Project.GetActiveProjectLanguage() == ProgrammingLanguages.CSharp) ? "csproj" : "vbproj";
 
-            var uwpProjectName = GenContext.ToolBox.Shell.GetActiveProjectName();
-            var uwpProjectFolder = GenContext.ToolBox.Shell.GetActiveProjectPath();
+            var uwpProjectName = GenContext.ToolBox.Shell.Project.GetActiveProjectName();
+            var uwpProjectFolder = GenContext.ToolBox.Shell.Project.GetActiveProjectPath();
             var uwpProjectPath = Path.Combine(uwpProjectFolder, $"{uwpProjectName}.{extension}");
 
             var coreProjectName = $"{uwpProjectName}.Core";
