+++ code/src/UI/VisualStudio/GenShell/VsGenShellVisualStudio.cs
@@ -1,0 +1,124 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using EnvDTE;
+using Microsoft.Build.Utilities;
+using Microsoft.Templates.Core.Gen.Shell;
+using Microsoft.Templates.UI.Threading;
+using Microsoft.VisualStudio.Setup.Configuration;
+using Microsoft.VisualStudio.Threading;
+
+namespace Microsoft.Templates.UI.VisualStudio.GenShell
+{
+    public class VsGenShellVisualStudio : IGenShellVisualStudio
+    {
+        private readonly AsyncLazy<DTE> _dte;
+
+        private readonly Lazy<ISetupInstance2> vsInstance = new Lazy<ISetupInstance2>(() =>
+        {
+            var setupConfiguration = new SetupConfiguration();
+            return setupConfiguration.GetInstanceForCurrentProcess() as ISetupInstance2;
+        });
+
+        private readonly List<string> installedPackageIds = new List<string>();
+
+        private string _vsVersionInstance = string.Empty;
+
+        private string _vsProductVersion = string.Empty;
+
+        public VsGenShellVisualStudio(AsyncLazy<DTE> dte)
+        {
+            _dte = dte;
+        }
+
+        public string GetVsVersion()
+        {
+            if (string.IsNullOrEmpty(_vsProductVersion))
+            {
+                ISetupConfiguration configuration = new SetupConfiguration() as ISetupConfiguration;
+                ISetupInstance instance = configuration.GetInstanceForCurrentProcess();
+                _vsProductVersion = instance.GetInstallationVersion();
+            }
+
+            return _vsProductVersion;
+        }
+
+        public string GetVsVersionAndInstance()
+        {
+            if (string.IsNullOrEmpty(_vsVersionInstance))
+            {
+                ISetupConfiguration configuration = new SetupConfiguration() as ISetupConfiguration;
+                ISetupInstance instance = configuration.GetInstanceForCurrentProcess();
+                string version = instance.GetInstallationVersion();
+                string instanceId = instance.GetInstanceId();
+                _vsVersionInstance = $"{version}-{instanceId}";
+            }
+
+            return _vsVersionInstance;
+        }
+
+        public bool IsBuildInProgress()
+        {
+            return SafeThreading.JoinableTaskFactory.Run(async () =>
+            {
+                return await IsBuildInProgressAsync();
+            });
+        }
+
+        public async Task<bool> IsBuildInProgressAsync()
+        {
+            await SafeThreading.JoinableTaskFactory.SwitchToMainThreadAsync();
+            var dte = await _dte.GetValueAsync();
+            return dte.Solution.SolutionBuild.BuildState == vsBuildState.vsBuildStateInProgress;
+        }
+
+        public bool IsDebuggerEnabled()
+        {
+            return SafeThreading.JoinableTaskFactory.Run(async () =>
+            {
+                return await IsDebuggerEnabledAsync();
+            });
+        }
+
+        public async Task<bool> IsDebuggerEnabledAsync()
+        {
+            await SafeThreading.JoinableTaskFactory.SwitchToMainThreadAsync();
+            var dte = await _dte.GetValueAsync();
+            return dte.Debugger.CurrentMode != dbgDebugMode.dbgDesignMode;
+        }
+
+        public bool IsSdkInstalled(string version)
+        {
+            var sdks = ToolLocationHelper.GetTargetPlatformSdks();
+
+            foreach (var sdk in sdks)
+            {
+                var versions = ToolLocationHelper.GetPlatformsForSDK(sdk.TargetPlatformIdentifier, sdk.TargetPlatformVersion);
+                if (versions.Any(v => v.Contains(version)))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        public List<string> GetInstalledPackageIds()
+        {
+            if (!installedPackageIds.Any())
+            {
+                foreach (var package in this.vsInstance.Value.GetPackages())
+                {
+                    installedPackageIds.Add(package.GetId());
+                }
+            }
+
+            return installedPackageIds;
+        }
+    }
+}
