+++ code/src/UI/VisualStudio/RightClickActions.cs
@@ -9,6 +9,7 @@ using System.Reflection;
 using Microsoft.Templates.Core;
 using Microsoft.Templates.Core.Diagnostics;
 using Microsoft.Templates.Core.Gen;
+using Microsoft.Templates.Core.Gen.Shell;
 using Microsoft.Templates.Core.Locations;
 using Microsoft.Templates.Core.PostActions.Catalog.Merge;
 using Microsoft.Templates.Core.Services;
@@ -16,6 +17,7 @@ using Microsoft.Templates.UI.Launcher;
 using Microsoft.Templates.UI.Resources;
 using Microsoft.Templates.UI.Services;
 using Microsoft.Templates.UI.Threading;
+using Microsoft.Templates.UI.VisualStudio.GenShell;
 using Microsoft.Templates.Utilities.Services;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.TemplateWizard;
@@ -35,7 +37,7 @@ namespace Microsoft.Templates.UI.VisualStudio
 
         private readonly GenerationService _generationService = GenerationService.Instance;
 
-        private static VsGenShell _shell;
+        private static IGenShell _shell;
 
         private const string BlankProjectType = "Blank";
 
@@ -69,13 +71,13 @@ namespace Microsoft.Templates.UI.VisualStudio
             try
             {
                 SetContext();
-                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Page, WizardTypeEnum.AddPage);
+                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.Project.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Page, WizardTypeEnum.AddPage);
                 var statusBarMessage = string.Format(StringRes.StatusBarNewItemAddPageSuccess, userSelection.Pages[0].Name);
                 FinishGeneration(userSelection, statusBarMessage);
             }
             catch (WizardBackoutException)
             {
-                _shell.ShowStatusBarMessage(StringRes.StatusBarNewItemAddPageCancelled);
+                _shell.UI.ShowStatusBarMessage(StringRes.StatusBarNewItemAddPageCancelled);
             }
         }
 
@@ -86,13 +88,13 @@ namespace Microsoft.Templates.UI.VisualStudio
             try
             {
                 SetContext();
-                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Feature, WizardTypeEnum.AddFeature);
+                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.Project.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Feature, WizardTypeEnum.AddFeature);
                 var statusBarMessage = string.Format(StringRes.StatusBarNewItemAddFeatureSuccess, userSelection.Features[0].Name);
                 FinishGeneration(userSelection, statusBarMessage);
             }
             catch (WizardBackoutException)
             {
-                _shell.ShowStatusBarMessage(StringRes.StatusBarNewItemAddFeatureCancelled);
+                _shell.UI.ShowStatusBarMessage(StringRes.StatusBarNewItemAddFeatureCancelled);
             }
         }
 
@@ -103,13 +105,13 @@ namespace Microsoft.Templates.UI.VisualStudio
             try
             {
                 SetContext();
-                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Service, WizardTypeEnum.AddService);
+                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.Project.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Service, WizardTypeEnum.AddService);
                 var statusBarMessage = string.Format(StringRes.StatusBarNewItemAddServiceSuccess, userSelection.Services[0].Name);
                 FinishGeneration(userSelection, statusBarMessage);
             }
             catch (WizardBackoutException)
             {
-                _shell.ShowStatusBarMessage(StringRes.StatusBarNewItemAddServiceCancelled);
+                _shell.UI.ShowStatusBarMessage(StringRes.StatusBarNewItemAddServiceCancelled);
             }
         }
 
@@ -120,20 +122,20 @@ namespace Microsoft.Templates.UI.VisualStudio
             try
             {
                 SetContext();
-                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Testing, WizardTypeEnum.AddTesting);
+                var userSelection = WizardLauncher.Instance.StartAddTemplate(_shell.Project.GetActiveProjectLanguage(), new VSStyleValuesProvider(), TemplateType.Testing, WizardTypeEnum.AddTesting);
                 var statusBarMessage = string.Format(StringRes.StatusBarNewItemAddTestingSuccess, userSelection.Testing[0].Name);
                 FinishGeneration(userSelection, statusBarMessage);
             }
             catch (WizardBackoutException)
             {
-                _shell.ShowStatusBarMessage(StringRes.StatusBarNewItemAddTestingCancelled);
+                _shell.UI.ShowStatusBarMessage(StringRes.StatusBarNewItemAddTestingCancelled);
             }
         }
 
         public bool Visible(TemplateType templateType)
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            if (!_shell.GetActiveProjectIsWts())
+            if (!_shell.Project.GetActiveProjectIsWts())
             {
                 return false;
             }
@@ -158,12 +160,12 @@ namespace Microsoft.Templates.UI.VisualStudio
         public bool Visible()
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            return _shell.GetActiveProjectIsWts();
+            return _shell.Project.GetActiveProjectIsWts();
         }
 
         public bool Enabled()
         {
-            return !_shell.IsDebuggerEnabled() && !_shell.IsBuildInProgress();
+            return !_shell.VisualStudio.IsDebuggerEnabled() && !_shell.VisualStudio.IsBuildInProgress();
         }
 
         public void OpenTempFolder()
@@ -186,14 +188,14 @@ namespace Microsoft.Templates.UI.VisualStudio
         {
             ThreadHelper.ThrowIfNotOnUIThread();
             EnsureGenContextInitialized();
-            var language = _shell.GetActiveProjectLanguage();
+            var language = _shell.Project.GetActiveProjectLanguage();
             if (GenContext.CurrentLanguage == language)
             {
                 GenContext.Current = this;
 
-                DestinationPath = GenContext.ToolBox.Shell.GetActiveProjectPath();
-                ProjectName = GenContext.ToolBox.Shell.GetActiveProjectName();
-                SafeProjectName = GenContext.ToolBox.Shell.GetActiveProjectNamespace();
+                DestinationPath = GenContext.ToolBox.Shell.Project.GetActiveProjectPath();
+                ProjectName = GenContext.ToolBox.Shell.Project.GetActiveProjectName();
+                SafeProjectName = GenContext.ToolBox.Shell.Project.GetActiveProjectNamespace();
 
                 GenerationOutputPath = GenContext.GetTempGenerationPath(ProjectName);
                 ProjectInfo = new ProjectInfo();
@@ -223,14 +225,14 @@ namespace Microsoft.Templates.UI.VisualStudio
 
             await SafeThreading.JoinableTaskFactory.SwitchToMainThreadAsync();
             _generationService.FinishGeneration(userSelection);
-            await _shell.ShowStatusBarMessageAsync(statusBarMessage);
+            await (_shell.UI as VsGenShellUI).ShowStatusBarMessageAsync(statusBarMessage);
         }
 
         private bool EnsureGenContextInitialized()
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            var projectLanguage = _shell.GetActiveProjectLanguage();
-            var projectPlatform = ProjectMetadataService.GetProjectMetadata(_shell.GetActiveProjectPath()).Platform;
+            var projectLanguage = _shell.Project.GetActiveProjectLanguage();
+            var projectPlatform = ProjectMetadataService.GetProjectMetadata(_shell.Project.GetActiveProjectPath()).Platform;
 
             if (!string.IsNullOrEmpty(projectLanguage) && !string.IsNullOrEmpty(projectPlatform))
             {
@@ -253,7 +255,7 @@ namespace Microsoft.Templates.UI.VisualStudio
         private static string GetTempGenerationFolder()
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-            string projectGuid = _shell.GetProjectGuidByName(_shell.GetActiveProjectName()).ToString();
+            string projectGuid = _shell.Project.GetProjectGuidByName(_shell.Project.GetActiveProjectName()).ToString();
             return Path.Combine(Path.GetTempPath(), Configuration.Current.TempGenerationFolderPath, projectGuid);
         }
